Correctifs pour les suggestions
