+++ app/Http/Controllers/Backend/Extras/SuggestionController.php
@@ -21,45 +21,6 @@ class SuggestionController extends Controller {
 
     }
 
-    public function all(Request $request){
-
-        // Récupération des suggestions à l'aide d'une ressources de modèle
-        $suggestions_data = Suggestion::select('*')->orderBy('created_at','desc')->get();
-
-        dump(Suggestion::all());
-        $suggestions_resources = (new SuggestionResources($suggestions_data));
-        $suggestions = $suggestions_resources->toArray($request);
-
-        $title = "Suggestions";
-
-        return view('app.extras.suggestions.all',[
-            'suggestions' => $suggestions,
-            'title' => $title
-        ]);
-        
-    }
-
-    public function my_suggestions(Request $request){
-
-        $user = UserAuth();
-
-        $suggestions_data = $user->suggestions;
-
-
-        dump($suggestions_data);
-
-        $suggestions_resources = (new SuggestionResources($suggestions_data));
-        $suggestions = $suggestions_resources->toArray($request);
-
-        $title = "Mes suggestions";
-
-        return view('app.extras.suggestions.my',[
-            'suggestions' => $suggestions,
-            'title' => $title
-        ]);
-        
-    }
-
     public function create_suggestion(Request $request)
     {
         if($request->filled('content')){
