+++ resources/views/app/feed/sidenav.blade.php
@@ -1,5 +1,5 @@
 <ul class="nav flex-column resac-nav-ul">
     <li class="nav-item nav-pills">
-        <a class='nav-link small' href="{{ route('backend.suggestions.all') }}">Suggestions</a>
+        <a class='nav-link small' href="{{ route('app.suggestions.all') }}">Suggestions</a>
     </li>
 </ul>
